Merging dev

+++ src/app/auth/_guards/role.guard.ts
@@ -14,25 +14,24 @@ export class RoleGuard implements CanActivate {
   isInternal = environment.internalAccess;
   constructor(private readonly router: Router, private readonly sessionService: SessionService) {}
   canActivate(route: ActivatedRouteSnapshot): Observable<boolean> | Promise<boolean> | boolean {
-    if (this.isInternal) {
-      if (sessionStorage.getItem('currentUser')) {
-        const expectedRole = route.data.expectedRole;
-        if (expectedRole === 'UHCI_Advocate') {
-          this.checkAdv = this.sessionService.checkAdvocateRole();
-          if (this.checkAdv.value) {
-            // logged in and advocate role, so return true
-            return true;
-          }
-        } else if (expectedRole === 'UHCI_Project') {
-          this.checkPro = this.sessionService.checkProjectRole();
-          this.checkExecutive = this.sessionService.checkExecutiveRole();
-          this.checkAdv = this.sessionService.checkAdvocateRole();
-          if (this.checkPro.value || this.checkExecutive.value || this.checkAdv.value) {
-            // logged in and project/ executive role so return true
-            return true;
-          }
+    if (this.isInternal && sessionStorage.getItem('currentUser')) {
+      const expectedRole = route.data.expectedRole;
+      if (expectedRole === 'UHCI_Advocate') {
+        this.checkAdv = this.sessionService.checkAdvocateRole();
+        if (this.checkAdv.value) {
+          // logged in and advocate role, so return true
+          return true;
+        }
+      } else if (expectedRole === 'UHCI_Project') {
+        this.checkPro = this.sessionService.checkProjectRole();
+        this.checkExecutive = this.sessionService.checkExecutiveRole();
+        this.checkAdv = this.sessionService.checkAdvocateRole();
+        if (this.checkPro.value || this.checkExecutive.value || this.checkAdv.value) {
+          // logged in and project/ executive role so return true
+          return true;
         }
       }
+
       // not logged in so redirect to login page with the return url
       sessionStorage.clear();
       sessionStorage.setItem('cache', JSON.stringify(false));
